+++ trace_generator/Point.h
@@ -14,7 +14,7 @@ class Point {
         int get_size();
         float dimension_value(int d);
         ~Point();
-        double distance(Point* p);
+        float distance(Point* p);
         Point* copy();
         friend std::ostream& operator<<(std::ostream& os, const Point& p);
